refactor(cli): extract hosted upload+finalize orchestration into a shared module (FDRS-656)
Move the best-effort presigned-PUT orchestration (events.jsonl, state
blobs, signals.jsonl) and the /finalize Score synthesis out of
runScenarioHosted.ts into src/hosted/uploadAndFinalize.ts so the new
`pome eval` command can reuse them without duplication.

Behavior-preserving for the hosted run path: warning text, the 30s PUT
timeout, empty-signals skip, and null-key fallbacks are unchanged;
redactJsonl moved verbatim. The runScenarioHosted test stub gains the
new HostedClient.createEvalSession member (unreachable from the runner).
